Add testing make targets.
The testing code that same has written should be run against clean repos and
DB. Hence there are make targets to create a testing config file as well as
the testing DB and repos that it points to. All of these are removed by the
eisting clean target.

The test suite will only run if there is something running on the IDE port,
but the makefile can't start a test run of the IDE autoamatically since this
would block that terminal. There is, however a test-run target which makes
this easy.